requirements.txt
asgiref==3.5.2
Django==4.0
django-ckeditor==6.5.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
